% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{free_currency_api_get_current}
\alias{free_currency_api_get_current}
\title{Current Rate}
\usage{
free_currency_api_get_current(currency_from, currency_to,
  compact = "ultra")
}
\arguments{
\item{currency_from}{The 3-letter ISO currency code}

\item{currency_to}{The 3-letter ISO currency code}

\item{compact}{The compactness of the output format. There are three options: y, n, ultra.}
}
\description{
https://free.currencyconverterapi.com/
API v6
API only allows 8 days at a time for free version.
Rate limit for free version is 100 requests per hour.
}
\details{
The Current Rate function returns a list of the exchange -- from currency and to currency --
and the current exchange rate as a numeric.
}
\examples{
\dontrun{
free_currency_api_get_current("USD", "PHP", "ultra")
}
}
