% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{free_currency_api_get_historical}
\alias{free_currency_api_get_historical}
\title{Historical Rates}
\usage{
free_currency_api_get_historical(currency_from, currency_to, start_date,
  end_date)
}
\arguments{
\item{currency_from}{The 3-letter ISO currency code}

\item{currency_to}{The 3-letter ISO currency code}

\item{start_date}{YYYY-mm-dd formatted date}

\item{end_date}{YYYY-mm-dd formatted date}
}
\description{
https://free.currencyconverterapi.com/
API v6
API only allows 8 days at a time for free version.
Rate limit for free version is 100 requests per hour.
}
\details{
The Historical Rates function returns a data.frame with 4 variables:
  date -- R date objects
  from -- the base currency
  to -- the exchange currency
  rate -- numeric representing the exchange value at that date
}
\examples{
\dontrun{
free_currency_api_get_historical("USD", "PHP", "2018-07-04", "2018-07-11")
}
}
