% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{free_currency_api_list_countries}
\alias{free_currency_api_list_countries}
\title{List Countries}
\usage{
free_currency_api_list_countries()
}
\description{
https://free.currencyconverterapi.com/api/v6/countries
}
\details{
List Currencies returns a list of all the two and three letter ISO codes for countries available,
along with the human readable versions of country and currency name, as well as the currency symbol.
}
