% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{free_currency_api_list_currencies}
\alias{free_currency_api_list_currencies}
\title{List Currencies}
\usage{
free_currency_api_list_currencies()
}
\description{
https://free.currencyconverterapi.com/api/v6/currencies
}
\details{
List Currencies returns a list of all the three letter ISO codes for currencies available,
along with their more human readable name.
}
